change stack to hold byte per dictionary value. pull & push from stack currently. upgrade find next end word to proper handle -1

+++ RiscVSimulator/Model/RiscVProgramResult.cs
@@ -10,11 +10,11 @@ namespace RiscVSimulator.Model
         public RiscVProgramResult()
         {
             Register = new Register[32];
-            Memory = new Dictionary<int, int>();
+            Memory = new Dictionary<int, byte>();
         }
         public Register[] Register { get; set; }
         public int StackTextFreePosition { get; set; }
         public int StackStaticDataFreePosition { get; set; }
-        public Dictionary<int, int> Memory { get; set; }
+        public Dictionary<int, byte> Memory { get; set; }
     }
 }
